feat: adds home page, adds login button to /
